updatehub: Move state_change_callback logic to firmware module
